Redux Routine: Make it more flexible allowing runtimes without middlewares
