+++ packages/redux-routine/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.0.0 (Unreleased)
+
+### Breaking change
+
+- This releases changes the public API of the package. We're now exporting `createRuntime` and `createMiddleware` APIs to allow more flexibility in using the co-routine runtime without using it as a Redux Middleware.
+
 ## 2.0.0 (2018-09-05)
 
 ### Breaking Change
